requirements.txt
Flask
requests
dialogflow
python-telegram-bot